backend/requirements.txt
pymongo >= 4.7.2
groq >= 0.5.0
transformers >=0.40.2
torch >=2.3.0
accelerate
google-cloud-aiplatform >= 1.38